Changed action button controller endpoint https://github.com/SoftUni-Internal/exam-systems-issues/issues/87
